+++ spec/fixtures/hiera/default.yaml
@@ -106,8 +106,8 @@ nebula::profile::vmhost::host::gateway: '10.1.2.3'
 nebula::profile::vmhost::host::nameservers: "%{alias('nebula::resolv_conf::nameservers')}"
 
 nebula::profile::named_instances::puma_wrapper: "/l/local/bin/profile_puma_wrap"
-nebula::profile::named_instances::fauxpaas_pubkey: "somepublickey"
-nebula::profile::named_instances::fauxpaas_puma_config: "config/fauxpaas_puma.rb"
+nebula::profile::named_instances::pubkey: "somepublickey"
+nebula::profile::named_instances::puma_config: "config/fauxpaas_puma.rb"
 
 nebula::profile::work_around_puppet_bugs::state_yaml_path: '/opt/puppetlabs/puppet/cache/state/state.yaml'
 nebula::profile::work_around_puppet_bugs::state_yaml_max_size: '10m'
